Added New Module qs_test_recorder_srt
This a new qs_test function that can be run at the command line level and from bash.  Arguments can be passed to this application.  The application will pass the test result to SynapseRT for the given test plan and test cycle and move files and/or zip directories and upload the zipped directory to SynapseRT with the test case.